+++ src/Describer/OpenApiPhpDescriber.php
@@ -14,12 +14,12 @@ namespace Nelmio\ApiDocBundle\Describer;
 use Nelmio\ApiDocBundle\Attribute\Operation;
 use Nelmio\ApiDocBundle\Attribute\Security;
 use Nelmio\ApiDocBundle\OpenApiPhp\Util;
+use Nelmio\ApiDocBundle\RouteDescriber\RouteDescriberTrait;
 use Nelmio\ApiDocBundle\Util\ControllerReflector;
 use Nelmio\ApiDocBundle\Util\SetsContextTrait;
 use OpenApi\Analysers\AttributeAnnotationFactory;
 use OpenApi\Annotations as OA;
 use OpenApi\Generator;
-use Psr\Log\LoggerInterface;
 use Symfony\Component\Routing\Route;
 use Symfony\Component\Routing\RouteCollection;
 
@@ -28,49 +28,59 @@ class_exists(OA\OpenApi::class);
 
 final class OpenApiPhpDescriber
 {
+    use RouteDescriberTrait;
     use SetsContextTrait;
 
-    private RouteCollection $routeCollection;
-    private ControllerReflector $controllerReflector;
-    private LoggerInterface $logger;
-
-    public function __construct(RouteCollection $routeCollection, ControllerReflector $controllerReflector, LoggerInterface $logger)
-    {
-        $this->routeCollection = $routeCollection;
-        $this->controllerReflector = $controllerReflector;
-        $this->logger = $logger;
+    public function __construct(
+        private readonly RouteCollection $routeCollection,
+        private readonly ControllerReflector $controllerReflector,
+        private readonly OperationIdGeneration $operationIdGeneration = OperationIdGeneration::ALWAYS_PREPEND,
+    ) {
     }
 
     public function describe(OA\OpenApi $api): void
     {
         $classAnnotations = [];
 
-        /** @var \ReflectionMethod $method */
-        foreach ($this->getMethodsToParse() as $method => [$path, $httpMethods, $routeName]) {
-            $declaringClass = $method->getDeclaringClass();
+        foreach ($this->getRoutesToParse() as $routeName => $route) {
+            $controller = $route->getDefault('_controller');
+            $reflectedMethod = $this->controllerReflector->getReflectionMethod($controller);
+            if (null === $reflectedMethod) {
+                continue;
+            }
+
+            $path = $this->normalizePath($route->getPath());
+            $supportedHttpMethods = $this->getSupportedHttpMethods($route);
+
+            $classReflector = $reflectedMethod->getDeclaringClass();
+            if (\is_array($controller) && method_exists(...$controller)) {
+                $classReflector = new \ReflectionClass($controller[0]);
+            } elseif (\is_string($controller) && false !== $i = strpos($controller, '::')) {
+                $classReflector = new \ReflectionClass(substr($controller, 0, $i));
+            }
 
             $path = Util::getPath($api, $path);
 
             $context = Util::createContext(['nested' => $path], $path->_context);
-            $context->namespace = $declaringClass->getNamespaceName();
-            $context->class = $declaringClass->getShortName();
-            $context->method = $method->name;
-            $context->filename = $method->getFileName();
+            $context->namespace = $classReflector->getNamespaceName();
+            $context->class = $classReflector->getShortName();
+            $context->method = $reflectedMethod->name;
+            $context->filename = $reflectedMethod->getFileName();
 
             $this->setContext($context);
 
-            if (!\array_key_exists($declaringClass->getName(), $classAnnotations)) {
-                $classAnnotations[$declaringClass->getName()] = $this->getAttributesAsAnnotation($declaringClass, $context);
+            if (!\array_key_exists($classReflector->getName(), $classAnnotations)) {
+                $classAnnotations[$classReflector->getName()] ??= $this->getAttributesAsAnnotation($classReflector, $context);
             }
 
-            $annotations = $this->getAttributesAsAnnotation($method, $context);
+            $annotations = $this->getAttributesAsAnnotation($reflectedMethod, $context);
 
             $implicitAnnotations = [];
             $mergeProperties = new \stdClass();
 
-            foreach (array_merge($annotations, $classAnnotations[$declaringClass->getName()]) as $annotation) {
+            foreach (array_merge($annotations, $classAnnotations[$classReflector->getName()]) as $annotation) {
                 if ($annotation instanceof Operation) {
-                    foreach ($httpMethods as $httpMethod) {
+                    foreach ($supportedHttpMethods as $httpMethod) {
                         $operation = Util::getOperation($path, $httpMethod);
                         $operation->mergeProperties($annotation);
                     }
@@ -79,7 +89,7 @@ final class OpenApiPhpDescriber
                 }
 
                 if ($annotation instanceof OA\Operation) {
-                    if (!\in_array($annotation->method, $httpMethods, true)) {
+                    if (!\in_array($annotation->method, $supportedHttpMethods, true)) {
                         continue;
                     }
                     if (Generator::UNDEFINED !== $annotation->path && $path->path !== $annotation->path) {
@@ -95,13 +105,21 @@ final class OpenApiPhpDescriber
                 if ($annotation instanceof Security) {
                     $annotation->validate();
 
-                    if (null === $annotation->name) {
-                        $mergeProperties->security = [];
+                    foreach ($supportedHttpMethods as $httpMethod) {
+                        $operation = Util::getOperation($path, $httpMethod);
 
-                        continue;
+                        if (Generator::UNDEFINED === $operation->security) {
+                            $operation->security = [];
+                        }
+
+                        if (null === $annotation->name) {
+                            $operation->security = [];
+
+                            continue;
+                        }
+
+                        $operation->security[] = [$annotation->name => $annotation->scopes];
                     }
-
-                    $mergeProperties->security[] = [$annotation->name => $annotation->scopes];
 
                     continue;
                 }
@@ -122,13 +140,13 @@ final class OpenApiPhpDescriber
                     && !$annotation instanceof OA\Parameter
                     && !$annotation instanceof OA\ExternalDocumentation
                 ) {
-                    throw new \LogicException(\sprintf('Using the annotation "%s" as a root annotation in "%s::%s()" is not allowed.', $annotation::class, $method->getDeclaringClass()->name, $method->name));
+                    throw new \LogicException(\sprintf('Using the annotation "%s" as a root annotation in "%s::%s()" is not allowed.', $annotation::class, $reflectedMethod->getDeclaringClass()->name, $reflectedMethod->name));
                 }
 
                 $implicitAnnotations[] = $annotation;
             }
 
-            foreach ($httpMethods as $httpMethod) {
+            foreach ($supportedHttpMethods as $httpMethod) {
                 $operation = Util::getOperation($path, $httpMethod);
                 if ([] !== $implicitAnnotations) {
                     $operation->merge($implicitAnnotations);
@@ -138,7 +156,11 @@ final class OpenApiPhpDescriber
                 }
 
                 if (Generator::UNDEFINED === $operation->operationId) {
-                    $operation->operationId = $httpMethod.'_'.$routeName;
+                    $operation->operationId = match ($this->operationIdGeneration) {
+                        OperationIdGeneration::ALWAYS_PREPEND => $httpMethod.'_'.$routeName,
+                        OperationIdGeneration::CONDITIONALLY_PREPEND => str_starts_with($routeName, $httpMethod) ? $routeName : $httpMethod.'_'.$routeName,
+                        OperationIdGeneration::NO_PREPEND => $routeName,
+                    };
                 }
             }
         }
@@ -147,28 +169,9 @@ final class OpenApiPhpDescriber
         $this->setContext(null);
     }
 
-    private function getMethodsToParse(): \Generator
+    private function getRoutesToParse(): \Generator
     {
-        foreach ($this->routeCollection->all() as $routeName => $route) {
-            if (!$route->hasDefault('_controller')) {
-                continue;
-            }
-            $controller = $route->getDefault('_controller');
-            $reflectedMethod = $this->controllerReflector->getReflectionMethod($controller);
-            if (null === $reflectedMethod) {
-                continue;
-            }
-            $path = $this->normalizePath($route->getPath());
-            $supportedHttpMethods = $this->getSupportedHttpMethods($route);
-            if ([] === $supportedHttpMethods) {
-                $this->logger->warning('None of the HTTP methods specified for path {path} are supported by swagger-ui, skipping this path', [
-                    'path' => $path,
-                ]);
-
-                continue;
-            }
-            yield $reflectedMethod => [$path, $supportedHttpMethods, $routeName];
-        }
+        yield from $this->routeCollection->all();
     }
 
     /**
@@ -185,15 +188,6 @@ final class OpenApiPhpDescriber
         }
 
         return array_intersect($methods, $allMethods);
-    }
-
-    private function normalizePath(string $path): string
-    {
-        if ('.{_format}' === substr($path, -10)) {
-            $path = substr($path, 0, -10);
-        }
-
-        return $path;
     }
 
     /**