+++ tests/Functional/ConfigurableContainerFactory.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Nelmio\ApiDocBundle\Tests\Functional;
 
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -25,16 +26,17 @@ final class ConfigurableContainerFactory
     private $container;
 
     /**
-     * @param Bundle[] $extraBundles
-     * @param string[] $extraConfigs
+     * @param Bundle[]                    $extraBundles
+     * @param array<string, array<mixed>> $extraConfigs     Key is the extension name, value is the config
+     * @param array<string, Definition>   $extraDefinitions
      */
-    public function create(array $extraBundles, ?callable $routeConfiguration, array $extraConfigs): void
+    public function create(array $extraBundles, ?callable $routeConfiguration, array $extraConfigs, array $extraDefinitions): void
     {
         // clear cache directory for a fresh container
         $filesystem = new Filesystem();
         $filesystem->remove('var/cache/test');
 
-        $appKernel = new NelmioKernel($extraBundles, $routeConfiguration, $extraConfigs);
+        $appKernel = new NelmioKernel($extraBundles, $routeConfiguration, $extraConfigs, $extraDefinitions);
         $appKernel->boot();
 
         $this->container = $appKernel->getContainer();