+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # CHANGELOG
 
+## 5.2.1
+Fixed a bug where using abstract controllers would ignore various attributes like ``#[OA\Tag]`` & ``#[Security]`` on child classes
+
+## 5.2.0
+Made it possible to automatically generate security definitions based on the ``#[IsGranted]`` attribute.
+
+```yaml
+nelmio_api_doc:
+    # ...
+
+    areas:
+        default:
+            security:
+                MyBearerScheme:
+                    type: 'http'
+                    scheme: 'bearer'
+```
+
+## 5.1.0
+Made it possible to configure how operation ids are generated. 
+
+```yaml
+nelmio_api_doc:
+    operation_id_generation: always_prepend
+```
+
+Possible values: ``always_prepend``, ``conditionally_prepend``, ``no_prepend`` or enum instance of ``Nelmio\ApiDocBundle\Describer\OperationIdGeneration``
+
 ## 4.38.2
 - Support of attribute MapQueryParameter with a regexp has been improved, it now converts the regexp from PCRE to ECMA-262 for better compliance with OpenApi.